pyproject.toml
[project]
name = "kagent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "torch_geometric",
    "einops",
    "timm",
    "rich",
    "python-dotenv",
    "simple-parsing",
    "wandb",
    "tqdm",
    "matplotlib",
    "pyyaml",
]

[tool.setuptools.packages.find]
exclude = ["k8s*", "instructions*", "sample_competition*"]

[project.optional-dependencies]
dev = [
    "ipython",
    "jupyter",
]
